+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.0.0
 
-- Contracts: complete stake requested event  ([#30](https://github.com/OpenSTFoundation/openst-tokens/pull/30))
-- Contracts: add requestStake  ([#30](https://github.com/OpenSTFoundation/openst-tokens/pull/30))
+- Contracts: complete EIP20 view functions ([#32](https://github.com/OpenSTFoundation/openst-tokens/pull/32))
+- Contracts: complete stake request accepted event ([#32](https://github.com/OpenSTFoundation/openst-tokens/pull/32))
+- Contracts: add acceptStakeRequest ([#32](https://github.com/OpenSTFoundation/openst-tokens/pull/32))
+- Contracts: complete stake requested event ([#30](https://github.com/OpenSTFoundation/openst-tokens/pull/30))
+- Contracts: add requestStake ([#30](https://github.com/OpenSTFoundation/openst-tokens/pull/30))
 - Contracts: add value token to constructor ([#25](https://github.com/OpenSTFoundation/openst-tokens/pull/25))
 - Contracts: inherit EIP20 interface ([#25](https://github.com/OpenSTFoundation/openst-tokens/pull/25))
 - Contracts: import EIP20 interface ([#25](https://github.com/OpenSTFoundation/openst-tokens/pull/25))